# src/backend/swagger.yaml
openapi: 3.0.3
info:
  title: Flashcard API Documentation
  description: API documentation for Flashcard application
  version: 1.0.0
servers:
  - url: http://localhost:3000/v1/api
  - url: https://fatcat-flashcard-app.onrender.com/v1/api
...
paths:
...
  /deck:
    get:
      tags:
        - deck
      summary: Get all decks
      description: Retrieve all decks
      operationId: getAllDecks
      responses:
        '200':
          description: List of all decks
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Get all decks successfully.
                  metadata:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deck'

    post:
      tags:
        - deck
      summary: Create new deck
      description: Create a new deck
      operationId: createDeck
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeckInput'
      responses:
        '200':
          description: Deck created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Create new deck successfully.
                  metadata:
                    $ref: '#/components/schemas/Deck'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Deck:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        user_id:
          type: string
        issuer_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DeckInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...